+++ pkg/sink/kafka/options_test.go
@@ -17,7 +17,6 @@ import (
 	"context"
 	"fmt"
 	"net/url"
-	"strconv"
 	"testing"
 	"time"
 
@@ -25,9 +24,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tiflow/cdc/contextutil"
 	"github.com/pingcap/tiflow/cdc/model"
-	"github.com/pingcap/tiflow/cdc/sink/codec/common"
-	"github.com/pingcap/tiflow/cdc/sink/mq/producer/kafka"
-	"github.com/pingcap/tiflow/pkg/config"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
 	"github.com/pingcap/tiflow/pkg/security"
 	"github.com/stretchr/testify/require"
@@ -196,233 +192,6 @@ func TestSetPartitionNum(t *testing.T) {
 	options.PartitionNum = 3
 	err = options.SetPartitionNum(2)
 	require.True(t, cerror.ErrKafkaInvalidPartitionNum.Equal(err))
-}
-
-func TestConfigurationCombinations(t *testing.T) {
-	kafka.NewAdminClientImpl = NewMockAdminClient
-	defer func() {
-		kafka.NewAdminClientImpl = NewSaramaAdminClient
-	}()
-
-	combinations := []struct {
-		uriTemplate             string
-		uriParams               []interface{}
-		brokerMessageMaxBytes   string
-		topicMaxMessageBytes    string
-		expectedMaxMessageBytes string
-	}{
-		// topic not created,
-		// `max-message-bytes` not set, `message.max.bytes` < `max-message-bytes`
-		// expected = min(`max-message-bytes`, `message.max.bytes`) = `message.max.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{"not-exist-topic"},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			BrokerMessageMaxBytes,
-		},
-		// topic not created,
-		// `max-message-bytes` not set, `message.max.bytes` = `max-message-bytes`
-		// expected = min(`max-message-bytes`, `message.max.bytes`) = `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{"not-exist-topic"},
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-			TopicMaxMessageBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-		},
-		// topic not created,
-		// `max-message-bytes` not set, broker `message.max.bytes` > `max-message-bytes`
-		// expected = min(`max-message-bytes`, `message.max.bytes`) = `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{"no-params"},
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			TopicMaxMessageBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-		},
-
-		// topic not created
-		// user set `max-message-bytes` < `message.max.bytes` < default `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(1024*1024 - 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			strconv.Itoa(1024*1024 - 1),
-		},
-		// topic not created
-		// user set `max-message-bytes` < default `max-message-bytes` < `message.max.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(config.DefaultMaxMessageBytes - 1)},
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			TopicMaxMessageBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes - 1),
-		},
-		// topic not created
-		// `message.max.bytes` < user set `max-message-bytes` < default `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(1024*1024 + 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			BrokerMessageMaxBytes,
-		},
-		// topic not created
-		// `message.max.bytes` < default `max-message-bytes` < user set `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(config.DefaultMaxMessageBytes + 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			BrokerMessageMaxBytes,
-		},
-		// topic not created
-		// default `max-message-bytes` < user set `max-message-bytes` < `message.max.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(config.DefaultMaxMessageBytes + 1)},
-			strconv.Itoa(config.DefaultMaxMessageBytes + 2),
-			TopicMaxMessageBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-		},
-		// topic not created
-		// default `max-message-bytes` < `message.max.bytes` < user set `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{"not-created-topic", strconv.Itoa(config.DefaultMaxMessageBytes + 2)},
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			TopicMaxMessageBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-		},
-
-		// topic created,
-		// `max-message-bytes` not set, topic's `max.message.bytes` < `max-message-bytes`
-		// expected = min(`max-message-bytes`, `max.message.bytes`) = `max.message.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{DefaultMockTopicName},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			TopicMaxMessageBytes,
-		},
-		// `max-message-bytes` not set, topic created, topic's `max.message.bytes` = `max-message-bytes`
-		// expected = min(`max-message-bytes`, `max.message.bytes`) = `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{DefaultMockTopicName},
-			BrokerMessageMaxBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-		},
-		// `max-message-bytes` not set, topic created, topic's `max.message.bytes` > `max-message-bytes`
-		// expected = min(`max-message-bytes`, `max.message.bytes`) = `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s",
-			[]interface{}{DefaultMockTopicName},
-			BrokerMessageMaxBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			strconv.Itoa(config.DefaultMaxMessageBytes),
-		},
-
-		// topic created
-		// user set `max-message-bytes` < `max.message.bytes` < default `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(1024*1024 - 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			strconv.Itoa(1024*1024 - 1),
-		},
-		// topic created
-		// user set `max-message-bytes` < default `max-message-bytes` < `max.message.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(config.DefaultMaxMessageBytes - 1)},
-			BrokerMessageMaxBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			strconv.Itoa(config.DefaultMaxMessageBytes - 1),
-		},
-		// topic created
-		// `max.message.bytes` < user set `max-message-bytes` < default `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(1024*1024 + 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			TopicMaxMessageBytes,
-		},
-		// topic created
-		// `max.message.bytes` < default `max-message-bytes` < user set `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(config.DefaultMaxMessageBytes + 1)},
-			BrokerMessageMaxBytes,
-			TopicMaxMessageBytes,
-			TopicMaxMessageBytes,
-		},
-		// topic created
-		// default `max-message-bytes` < user set `max-message-bytes` < `max.message.bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(config.DefaultMaxMessageBytes + 1)},
-			BrokerMessageMaxBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 2),
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-		},
-		// topic created
-		// default `max-message-bytes` < `max.message.bytes` < user set `max-message-bytes`
-		{
-			"kafka://127.0.0.1:9092/%s?max-message-bytes=%s",
-			[]interface{}{DefaultMockTopicName, strconv.Itoa(config.DefaultMaxMessageBytes + 2)},
-			BrokerMessageMaxBytes,
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-			strconv.Itoa(config.DefaultMaxMessageBytes + 1),
-		},
-	}
-
-	for _, a := range combinations {
-		BrokerMessageMaxBytes = a.brokerMessageMaxBytes
-		TopicMaxMessageBytes = a.topicMaxMessageBytes
-
-		uri := fmt.Sprintf(a.uriTemplate, a.uriParams...)
-		sinkURI, err := url.Parse(uri)
-		require.Nil(t, err)
-
-		options := NewOptions()
-		err = options.Apply(sinkURI)
-		require.Nil(t, err)
-
-		saramaConfig, err := NewSaramaConfig(context.Background(), options)
-		require.Nil(t, err)
-
-		adminClient, err := kafka.NewAdminClientImpl([]string{sinkURI.Host}, saramaConfig)
-		require.Nil(t, err)
-
-		topic, ok := a.uriParams[0].(string)
-		require.True(t, ok)
-		require.NotEqual(t, "", topic)
-		err = kafka.AdjustConfig(adminClient, options, saramaConfig, topic)
-		require.Nil(t, err)
-
-		encoderConfig := common.NewConfig(config.ProtocolOpen)
-		err = encoderConfig.Apply(sinkURI, &config.ReplicaConfig{})
-		require.Nil(t, err)
-		encoderConfig.WithMaxMessageBytes(saramaConfig.Producer.MaxMessageBytes)
-
-		err = encoderConfig.Validate()
-		require.Nil(t, err)
-
-		// producer's `MaxMessageBytes` = encoder's `MaxMessageBytes`.
-		require.Equal(t, encoderConfig.MaxMessageBytes, saramaConfig.Producer.MaxMessageBytes)
-
-		expected, err := strconv.Atoi(a.expectedMaxMessageBytes)
-		require.Nil(t, err)
-		require.Equal(t, expected, saramaConfig.Producer.MaxMessageBytes)
-
-		_ = adminClient.Close()
-	}
 }
 
 func TestApplySASL(t *testing.T) {
